made a new branch for unstable releases
I also went ahead and deleted comments I didn't find ver useful or needed so I just deleted them. I also found out in the discord.py documentation that instead of using an if statement to see if the invoke message was used in a guild, I could just used the @guild_only flag and it would do it for me; saving me a lot of time.